requirements.txt
Flask
requests
python-dotenv
fastapi
uvicorn

